+++ src/components/package/_fields/visibility-edit/visibility-edit.js
@@ -1,0 +1,43 @@
+import { FormattedMessage } from 'react-intl';
+import {
+  Field,
+} from 'react-final-form';
+import { FieldArray } from 'react-final-form-arrays';
+
+import {
+  Headline,
+  Checkbox,
+} from '@folio/stripes/components';
+
+import fieldsetStyles from '../../../fieldset-styles.css';
+
+export const VisibilityEdit = () => {
+  const renderFieldGroup = ({ fields }) => (
+    <>
+      {fields.value.map(({ category }, index) => (
+        <Field
+          key={category}
+          type="checkbox"
+          component={Checkbox}
+          label={<FormattedMessage id={`ui-eholdings.package.visibility.${category}`} />}
+          name={`visibility[${index}].hidden`}
+        />
+      ))}
+    </>
+  );
+
+  return (
+    <>
+      <Headline
+        tag="legend"
+        className={fieldsetStyles.label}
+      >
+        <FormattedMessage id="ui-eholdings.package.visibility" />
+      </Headline>
+      <FieldArray
+        name="visibility"
+        render={renderFieldGroup}
+      />
+    </>
+  );
+};