+++ src/components/package/edit-managed/managed-package-edit.test.js
@@ -28,10 +28,11 @@ const model = {
   packageToken: {
     value: 'token',
   },
-  visibilityData: {
-    isHidden: false,
-    reason: '',
-  },
+  visibility: [
+    { category: 'PF', hidden: false, reason: '' },
+    { category: 'FTF', hidden: false, reason: '' },
+    { category: 'MARC', hidden: false, reason: '' },
+  ],
   proxy: {
     id: 'proxy-id',
     isInherited: true,