+++ src/components/package/show/components/package-settings/fields/visibility-view.js
@@ -1,0 +1,19 @@
+import { FormattedMessage } from 'react-intl';
+
+import {
+  Checkbox,
+  KeyValue,
+} from '@folio/stripes/components';
+
+export const VisibilityView = ({ visibility }) => (
+  <KeyValue label={<FormattedMessage id="ui-eholdings.package.visibility" />}>
+    {visibility.map(visibilityOption => (
+      <Checkbox
+        key={visibilityOption.category}
+        label={<FormattedMessage id={`ui-eholdings.package.visibility.${visibilityOption.category}`} />}
+        checked={visibilityOption.hidden}
+        disabled
+      />
+    ))}
+  </KeyValue>
+);